+++ spnkr/services/base.py
@@ -13,39 +13,24 @@ Response: TypeAlias = "ClientResponse | CachedResponse"
 Session: TypeAlias = "ClientSession | CachedSession"
 
 
-def _create_limiter(rate_per_second: int) -> AsyncLimiter:
-    """Return an AsyncLimiter with the given rate per second."""
-    # Setting the max rate to 1 disallows bursts.
-    return AsyncLimiter(1, 1 / rate_per_second)
-
-
-def _is_cached_response(response: Response) -> bool:
-    """Return `True` if the response is a cached response."""
-    return hasattr(response, "from_cache")
-
-
 class BaseService:
     """Base service class. Handles initialization and rate limiting requests."""
 
-    def __init__(
-        self, session: Session, requests_per_second: int | None = 5
-    ) -> None:
+    def __init__(self, session: Session, requests_per_second: int = 5) -> None:
         """Initialize a service.
 
         Args:
             session: The authenticated aiohttp session to use.
-            requests_per_second: The rate limit to use. Set to None to disable
-                rate limiting.
+            requests_per_second: The rate limit to use.
         """
         self._session = session
-        self._rate_limiter = None
-        if requests_per_second is not None:
-            self._rate_limiter = _create_limiter(requests_per_second)
+        # Setting the max rate to 1 disallows bursts.
+        self._rate_limiter = AsyncLimiter(1, 1 / requests_per_second)
 
     async def _get(self, url: str, **kwargs) -> Response:
         """Make a GET request to `url` and return the response."""
         response = await self._session.get(url, **kwargs)
-        if not _is_cached_response(response) and self._rate_limiter is not None:
+        if not hasattr(response, "from_cache"):
             # Only rate limit non-cached responses.
             await self._rate_limiter.acquire()
         response.raise_for_status()