+++ CHANGES.md
@@ -2,9 +2,11 @@
 
 ## [Unreleased]
 
+## [0.9.0] - 2024-08-26
+
 ### Added
 
-- Read functionality for film chunks.
+- Read functionality for highlight event film chunks.
 
 ### Changed
 
@@ -123,7 +125,8 @@
 
 First documented release.
 
-[unreleased]: https://github.com/acurtis166/SPNKr/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/acurtis166/SPNKr/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/acurtis166/SPNKr/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/acurtis166/SPNKr/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/acurtis166/SPNKr/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/acurtis166/SPNKr/compare/v0.5.0...v0.6.0